+++ packages/personal/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @stellar-agent-kit/personal
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @stellar-agent-kit/core@0.1.4
+
 ## 0.1.3
 
 ### Patch Changes